requirements.txt
gunicorn==20.0.4
pandas==1.0.3
uvicorn==0.11.2
PyYAML==5.3
mock==4.0.1
starlette==0.12.9
fastapi==0.48.0
requests==2.23.0
redis==3.4.1
xlrd==1.2.0
orjson==2.6.1
